fixed #924
