+++ src/common/api/moduleApi.type.ts
@@ -5,15 +5,10 @@
 // that can be found in the LICENSE file exposed on Github (readium) in the project repository.
 // ==LICENSE-END==
 
-import { TMethodApi } from "readium-desktop/common/api/methodApi.type";
-import { TModuleApi } from "readium-desktop/common/api/moduleApi.type";
-
-export interface MetaApi {
-    requestId: string;
-    moduleId: TModuleApi;
-    methodId: TMethodApi;
-}
-
-export interface Meta {
-    api: MetaApi;
-}
+// module typing
+type TCatalogApi = "catalog";
+type TPublicationApi = "publication";
+type TOpdsApi = "opds";
+type TLcpApi = "lcp";
+type TReaderApi = "reader";
+export type TModuleApi = TCatalogApi | TPublicationApi | TOpdsApi | TLcpApi | TReaderApi;
