+++ src/common/redux/actions/api/request.ts
@@ -5,8 +5,9 @@
 // that can be found in the LICENSE file exposed on Github (readium) in the project repository.
 // ==LICENSE-END==
 
+import { TMethodApi } from "readium-desktop/common/api/methodApi.type";
+import { TModuleApi } from "readium-desktop/common/api/moduleApi.type";
 import { Action } from "readium-desktop/common/models/redux";
-import { TMethodApi, TModuleApi } from "readium-desktop/main/di";
 
 import { Meta } from "./types";
