+++ src/main/api/lcp.ts
@@ -7,6 +7,7 @@
 
 import * as debug_ from "debug";
 import { inject, injectable } from "inversify";
+import { ILcpApi } from "readium-desktop/common/api/interface/lcpApi.interface";
 import { lcpActions } from "readium-desktop/common/redux/actions";
 import { readerActions } from "readium-desktop/common/redux/actions/";
 import { Translator } from "readium-desktop/common/services/translator";
@@ -19,18 +20,6 @@ import { PublicationViewConverter } from "../converter/publication";
 import { RootState } from "../redux/states";
 
 const debug = debug_("readium-desktop:main:redux:sagas:streamer");
-
-export interface ILcpApi {
-    renewPublicationLicense: (publicationIdentifier: string) => Promise<void>;
-    returnPublication: (publicationIdentifier: string) => Promise<void>;
-    unlockPublicationWithPassphrase: (passphrase: string, publicationViewIdentifer: string) => Promise<void>;
-}
-
-export interface ILcpModuleApi {
-    "lcp/renewPublicationLicense": ILcpApi["renewPublicationLicense"];
-    "lcp/returnPublication": ILcpApi["returnPublication"];
-    "lcp/unlockPublicationWithPassphrase": ILcpApi["unlockPublicationWithPassphrase"];
-}
 
 @injectable()
 export class LcpApi implements ILcpApi {
