+++ src/renderer/common/redux/sagas/api.ts
@@ -5,9 +5,10 @@
 // that can be found in the LICENSE file exposed on Github (readium) in the project repository.
 // ==LICENSE-END==
 
+import { TApiMethod, TApiMethodName } from "readium-desktop/common/api/api.type";
+import { TMethodApi } from "readium-desktop/common/api/methodApi.type";
+import { TModuleApi } from "readium-desktop/common/api/moduleApi.type";
 import { apiActions } from "readium-desktop/common/redux/actions";
-import { TApiMethod, TApiMethodName } from "readium-desktop/main/api/api.type";
-import { TMethodApi, TModuleApi } from "readium-desktop/main/di";
 import { put } from "redux-saga/effects";
 
 /**
